+++ beacon-chain/types/crystallized_state_test.go
@@ -66,8 +66,8 @@ func TestInitialDeriveCrystallizedState(t *testing.T) {
 		t.Fatalf("expected justified streak to equal %d: got %d", 0, newCState.JustifiedStreak())
 	}
 
-	if newCState.LastStateRecalc() != uint64(params.CycleLength) {
-		t.Fatalf("expected last state recalc to equal %d: got %d", uint64(params.CycleLength), newCState.LastStateRecalc())
+	if newCState.LastStateRecalc() != params.CycleLength {
+		t.Fatalf("expected last state recalc to equal %d: got %d", params.CycleLength, newCState.LastStateRecalc())
 	}
 
 	if newCState.LastFinalizedSlot() != 0 {
@@ -96,7 +96,7 @@ func TestNextDeriveCrystallizedSlot(t *testing.T) {
 	}
 
 	totalDeposits := cState.TotalDeposits()
-	recentBlockHashes := make([][]byte, 3*uint64(params.CycleLength))
+	recentBlockHashes := make([][]byte, 3*params.CycleLength)
 	voteCache := make(map[[32]byte]*VoteCache)
 	for i := 0; i < 3*int(params.CycleLength); i++ {
 		blockHash := [32]byte{}
@@ -116,14 +116,14 @@ func TestNextDeriveCrystallizedSlot(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to derive crystallized state: %v", err)
 	}
-	if cState.LastStateRecalc() != 2*uint64(params.CycleLength) {
-		t.Fatalf("expected last state recalc to equal %d: got %d", 2*uint64(params.CycleLength), cState.LastStateRecalc())
+	if cState.LastStateRecalc() != 2*params.CycleLength {
+		t.Fatalf("expected last state recalc to equal %d: got %d", 2*params.CycleLength, cState.LastStateRecalc())
 	}
-	if cState.LastJustifiedSlot() != uint64(params.CycleLength)-1 {
-		t.Fatalf("expected justified slot to equal %d: got %d", uint64(params.CycleLength)-1, cState.LastJustifiedSlot())
+	if cState.LastJustifiedSlot() != params.CycleLength-1 {
+		t.Fatalf("expected justified slot to equal %d: got %d", params.CycleLength-1, cState.LastJustifiedSlot())
 	}
-	if cState.JustifiedStreak() != 2*uint64(params.CycleLength) {
-		t.Fatalf("expected justified streak to equal %d: got %d", uint64(params.CycleLength), cState.JustifiedStreak())
+	if cState.JustifiedStreak() != 2*params.CycleLength {
+		t.Fatalf("expected justified streak to equal %d: got %d", params.CycleLength, cState.JustifiedStreak())
 	}
 	if cState.LastFinalizedSlot() != 0 {
 		t.Fatalf("expected finalized slot to equal %d: got %d", 0, cState.LastFinalizedSlot())
@@ -133,17 +133,17 @@ func TestNextDeriveCrystallizedSlot(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to derive crystallized state: %v", err)
 	}
-	if cState.LastStateRecalc() != 3*uint64(params.CycleLength) {
-		t.Fatalf("expected last state recalc to equal %d: got %d", 3*uint64(params.CycleLength), cState.LastStateRecalc())
+	if cState.LastStateRecalc() != 3*params.CycleLength {
+		t.Fatalf("expected last state recalc to equal %d: got %d", 3*params.CycleLength, cState.LastStateRecalc())
 	}
-	if cState.LastJustifiedSlot() != 2*uint64(params.CycleLength)-1 {
-		t.Fatalf("expected justified slot to equal %d: got %d", 2*uint64(params.CycleLength)-1, cState.LastJustifiedSlot())
+	if cState.LastJustifiedSlot() != 2*params.CycleLength-1 {
+		t.Fatalf("expected justified slot to equal %d: got %d", 2*params.CycleLength-1, cState.LastJustifiedSlot())
 	}
-	if cState.JustifiedStreak() != 3*uint64(params.CycleLength) {
-		t.Fatalf("expected justified streak to equal %d: got %d", 2*uint64(params.CycleLength), cState.JustifiedStreak())
+	if cState.JustifiedStreak() != 3*params.CycleLength {
+		t.Fatalf("expected justified streak to equal %d: got %d", 2*params.CycleLength, cState.JustifiedStreak())
 	}
-	if cState.LastFinalizedSlot() != uint64(params.CycleLength)-1 {
-		t.Fatalf("expected finalized slot to equal %d: got %d", uint64(params.CycleLength)-1, cState.LastFinalizedSlot())
+	if cState.LastFinalizedSlot() != params.CycleLength-1 {
+		t.Fatalf("expected finalized slot to equal %d: got %d", params.CycleLength-1, cState.LastFinalizedSlot())
 	}
 }
